Removes ttypescript and submits to badness.
Unfortunately, it seems that even after what I thought was a very compelling argument for transforming `.ts` to `.js`, the TypeScript compiler team is still adamantly holding their position that they should not do it.  `ttypescript` is not really maintained anymore, and doesn't work with file watching, and alternatives come with large numbers of transitive dependencies.  Because of this, I'm finally giving up on this fight and accepting that this part of the world will be bad for the foreseeable future and I should dedicate my time to fighting other battles.