Changes `useAsyncState` to not trigger component rerender.
By referring to `captureContainer.value` we were making it so any component that called `waitFor` or `reset` would get re-rendered anytime either of those functions were called in the future.
By changing to `.peek()`, we can avoid having the component subscribe to changes to the `captureContainer` when those methods are called.