Adds error ensuring helper.
Adds pure helper function that will take in an `unknown` and give back an `Error`.  Useful in `catch` clauses when you want to deal with an `Error` and expect everything to be an `Error` but can't be sure.

Also adds functions that sanely serialize `bigint` and `Uint8Array` to JSON (used in `ensureError(...)`.